pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "soumyagpt"
version = "0.1.0"
description = "Voice-cloned CLI chatbot using Gemini API"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "Soumya", email = "your-email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    "click>=8.0.0",
    "faster-whisper>=1.0.0",
    "google-generativeai>=0.8.0",
    "pyttsx3>=2.90",
    "edge-tts>=6.1.0",
    "sounddevice>=0.4.0",
    "soundfile>=0.12.0",
    "numpy>=1.21.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "pydub>=0.25.0",
    "colorama>=0.4.0",
    "rich>=13.0.0",
    "scipy>=1.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
]

[project.scripts]
soumyagpt = "soumyagpt.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/soumyagpt"
Repository = "https://github.com/yourusername/soumyagpt.git"
Issues = "https://github.com/yourusername/soumyagpt/issues"

[tool.setuptools]
packages = ["soumyagpt"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true